Add Dip::Config::ConfigKeyMissingError
This commit adds the error class ConfigKeyMissingError which is used to indicate a config key is missing.

If a developer tries to access a config key that has a `nil` value, the ConfigKeyMissingError is raised with a description of which config is missing and the path to the config file.

Currently, when a config key is `nil` an error message such as "NoMethodError: undefined method `each_' for nil:NilClass" is raised.
This is an unhelpful message because the actual problem is implicit and it requires some cognitive load decipher.

By raising ConfigKeyMissingError when a config key is `nil`, developers are given explicit information about the error.